Overhaul selection mechanism
Now, clicking on a feature that is already selected will deselect it, while clicking on an unselected feature will always select it, even if another feature is already selected. Additionally, I have fixed a bug that was causing the selected feature style to not be removed when changing the selection from a linestring feature to a polygon feature.